+++ lib/features/inventory/domain/repositories/product_repository.dart
@@ -9,5 +9,5 @@ abstract class ProductRepository {
     List<VariantProductEntity> variants,
   );
   Future<void> deleteBaseProduct(int id);
-  Future<void> updateStock(int variantId, int delta, bool deductMp);
+  Future<void> updateStock(int variantId, double delta, bool deductMp);
 }
